[Gold II] Title: 가장 긴 증가하는 부분 수열 3, Time: 192 ms, Memory: 8292 KB -BaekjoonHub

+++ 백준/Gold/12738. 가장 긴 증가하는 부분 수열 3/가장 긴 증가하는 부분 수열 3.cc
@@ -2,8 +2,8 @@
 using namespace std;
 
 void solve(int n = 0) {
-  vector<long long> arr;
-  long long input;
+  vector<int> arr;
+  int input;
   while (n--) {
     cin >> input;
